Pull enum TollFreeVehicles up
