+++ Java/tollcalculator/TollFreeVehicles.java
@@ -1,0 +1,19 @@
+package tollcalculator;
+
+enum TollFreeVehicles {
+    MOTORBIKE("tollcalculator.Motorbike"),
+    TRACTOR("Tractor"),
+    EMERGENCY("Emergency"),
+    DIPLOMAT("Diplomat"),
+    FOREIGN("Foreign"),
+    MILITARY("Military");
+    private final String type;
+
+    TollFreeVehicles(String type) {
+        this.type = type;
+    }
+
+    public String getType() {
+        return type;
+    }
+}
